refactor: 이미지를 supabase에 업로드하는 기능 파일로 분리 (#147)
